created github flavour and mode error handler
